accuchek_reports: add a new driver for Accu-Chek Mobile meters.
The default USB connection provides a USB storage device with a CSV
datafile. This driver works by looking for that file and importing it.

This does not support setting the time, nor reading the actual device
time, but it is at least a good way to import data from different meters
under the same format.