fsinsulinx: add untested FreeStyle InsuLinx driver.
This is based off the fsprecisionneo driver, as the two share an almost
identical protocol. The readings format is different between them, but
otherwise the commands are compatible.

There are a number of higher-level commands that differ between them,
though, so it might be necessary to specialize the drivers further in the
future.