Abbott devices: add support for using hidapi library.
This allows these drivers to be used on operating systems other than Linux,
as hidapi abstracts away the access. The interface is similar enough to
using hidraw that the actual code only needed minimal changes.

Maintain the ability to access the devices with the hidraw path too,
without forcing the dependency.