serial devices: disable xonxoff by default.
With 2a825fb889735fa881566d1764cc48d2814447d2 the parameters to open the
serial device were lifted from the fsoptium driver, which was the only one
passing xonxoff=True.

The Optium device has no problem with disabling this feature, but the
codefree driver hangs if this is set to True. So instead set it to False
and get rid of it.