lifescan_binary_protocol: create a new module to support LifeScan drivers.
Both the UltraEasy and Verio 2015 use a similar protocol, with the same
base packet serialised to the device. Factor the packet definition out (and
make it a bit more generic). Verio IQ (Issue #30) also shares the same base
protocol.

Also move the definition of VERIO_TIMESTAMP to this common module as it's
also shared with the Verio IQ.