Fix for non-integer errors
This is a quick fix I used to address an error for "HI" readings on my unit.  This comes up when testing inhuman blood.  There may be a "LO", but I have not encountered it, so I do not know how the specifics of it.  This error may come up on other units, but I have just done something about the hardware I have access to.