Replace epydoc metadata with SPDX identifiers.
Most of the epydoc metadata (author, copyright, email) is duplicate in
the package's setup.py or README. These copies are more likely to get out
of date, and increase the amount of work to keep them in line.

The license information should rather be provided with an SPDX identifier,
which is more similar to what the Linux kernel and other projects, such as
pyserial, do.