FreeStyle devices: don't error out for non-ASCII patient names.
This was brought up in issue #60. While the devices don't appear to care
much for patient names, and can't be expected to support UTF-8 (the Windows
software don't appear to allow that), it's worth protecting against this
and just replacing the non-ASCII characters.

To make this easier, the decoding is pushed down to the end of the
processing of the multi-packet response, rather than being done per-packet.