Add a tool to extract a FreeStyle protocol chatter from an usbmon capture.
The tool needs https://github.com/Flameeyes/usbmon-tools to work, and is
derived off the extract-hid-chatter tool in that repository, so it's
Apache-2 licensed (unlike the rest of the repository) and shares copyright
with the usbmon-tools authors.

While this does not work correctly for Libre2 traces, it does appear to
work fine with Libre1 systems with text protocol exchanges, and it provides
a starting point.