Now that Python 3.6 is the minimum supported version, use f-strings.
The f-strings formatted strings are generally simpler to write, read and
debug, with a few exceptions.

Keep the logging strings lazy, keep the CSV-building with %-formatting, as
it makes it more readable.

In fsoptium, keep `'$%s\r\n'` as a %-format, to avoid making it
`f'${command}\r\n'` that would make it easy to mistake for a Shell variable.